requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
sqlparse==0.4.4
